Some metrics included
Now those metrics are there, at first the file system didnt work so i put it all into one file.

numbers are prob wrong

wasnt sure how to implement idle state for the three tellers like when to start the count for the three tellers and when to stop it, wasnt sure where to put the count incrementer.

and boom i attempted.